chore(CodebaseRefactor): Refactor code - Improve validations [Delivers #158139327]
